Sooooo i resolved the problem with the second audio, i added sourceNode= null; in the increaseStage() function. this way the audios will be initialized properly.
I worked my way through 6 stages. currently we are stuck on stage 6 where
the lottie animation with the dots is being played.
Several stages will be added soon. there is still task 3,
task 4(another lottieanimation) and the endscreen(loosingscreen) and
ofcourse the abortscreen(winningscreen) to code.

We also have to better the transitions between the stages.
the chooseanimal screen has to change too. there will be animals to choose inside
the circles and also sounds of the animal when selected.
Another thing to code is the saving of the animal till the end.